Creación de uiState para el manejo de estado, viewmodel para administrar estado. Reubicar la logica y los datos de la app en ViewModel, Llamar al viewModel
